More modular installation system for files.
All files are stored in a subfolder of the files directory. When
installing, this directory is searched recursively for folders
containing a .prefix file. The contents of this file are prepended
to the basename of every other file in the directory, and a symlink
is created from the result to the file in this repository.

Eventually, will want to do the same with folders.